Move become plugin tests to their proper directory and rename them accordingly
