+++ test/units/plugins/become/test_ksu.py
@@ -10,7 +10,7 @@ __metaclass__ = type
 import re
 import pytest
 
-from .test_play_context import parser, reset_cli_args
+from ...playbook.test_play_context import parser, reset_cli_args
 from ansible import constants as C
 from ansible import context
 from ansible.cli.arguments import option_helpers as opt_help
@@ -21,15 +21,15 @@ from ansible.plugins.loader import become_loader
 from ansible.utils import context_objects as co
 
 
-def test_play_context_make_become_cmd_dzdo(mocker, parser, reset_cli_args):
+def test_ksu(mocker, parser, reset_cli_args):
     options = parser.parse_args([])
     context._init_global_context(options)
     play_context = PlayContext()
 
     default_cmd = "/bin/foo"
     default_exe = "/bin/bash"
-    dzdo_exe = 'dzdo'
-    dzdo_flags = ''
+    ksu_exe = 'ksu'
+    ksu_flags = ''
 
     cmd = play_context.make_become_cmd(cmd=default_cmd, executable=default_exe)
     assert cmd == default_cmd
@@ -38,14 +38,9 @@ def test_play_context_make_become_cmd_dzdo(mocker, parser, reset_cli_args):
 
     play_context.become = True
     play_context.become_user = 'foo'
-    play_context.set_become_plugin(become_loader.get('dzdo'))
-    play_context.become_method = 'dzdo'
-    play_context.become_flags = dzdo_flags
+    play_context.set_become_plugin(become_loader.get('ksu'))
+    play_context.become_method = 'ksu'
+    play_context.become_flags = ksu_flags
     cmd = play_context.make_become_cmd(cmd=default_cmd, executable=default_exe)
-    assert re.match("""%s %s -u %s %s -c 'echo %s; %s'""" % (dzdo_exe, dzdo_flags, play_context.become_user, default_exe,
-                                                             success, default_cmd), cmd) is not None
-    play_context.become_pass = 'testpass'
-    play_context.set_become_plugin(become_loader.get('dzdo'))
-    cmd = play_context.make_become_cmd(cmd=default_cmd, executable=default_exe)
-    assert re.match("""%s %s -p %s -u %s %s -c 'echo %s; %s'""" % (dzdo_exe, dzdo_flags, r'\"\[dzdo via ansible, key=.+?\] password:\"',
-                                                                   play_context.become_user, default_exe, success, default_cmd), cmd) is not None
+    assert (re.match("""%s %s %s -e %s -c 'echo %s; %s'""" % (ksu_exe, play_context.become_user, ksu_flags,
+                                                              default_exe, success, default_cmd), cmd) is not None)
