+++ test/units/plugins/become/test_pbrun.py
@@ -10,7 +10,7 @@ __metaclass__ = type
 import re
 import pytest
 
-from .test_play_context import parser, reset_cli_args
+from ...playbook.test_play_context import parser, reset_cli_args
 from ansible import constants as C
 from ansible import context
 from ansible.cli.arguments import option_helpers as opt_help
@@ -21,15 +21,15 @@ from ansible.plugins.loader import become_loader
 from ansible.utils import context_objects as co
 
 
-def test_play_context_make_become_cmd_doas(mocker, parser, reset_cli_args):
+def test_pbrun(mocker, parser, reset_cli_args):
     options = parser.parse_args([])
     context._init_global_context(options)
     play_context = PlayContext()
 
     default_cmd = "/bin/foo"
     default_exe = "/bin/bash"
-    doas_exe = 'doas'
-    doas_flags = '-n'
+    pbrun_exe = 'pbrun'
+    pbrun_flags = ''
 
     cmd = play_context.make_become_cmd(cmd=default_cmd, executable=default_exe)
     assert cmd == default_cmd
@@ -38,9 +38,9 @@ def test_play_context_make_become_cmd_doas(mocker, parser, reset_cli_args):
 
     play_context.become = True
     play_context.become_user = 'foo'
-    play_context.set_become_plugin(become_loader.get('doas'))
-    play_context.become_method = 'doas'
-    play_context.become_flags = doas_flags
+    play_context.set_become_plugin(become_loader.get('pbrun'))
+    play_context.become_method = 'pbrun'
+    play_context.become_flags = pbrun_flags
     cmd = play_context.make_become_cmd(cmd=default_cmd, executable=default_exe)
-    assert (re.match("""%s %s -u %s %s -c 'echo %s; %s'""" % (doas_exe, doas_flags, play_context.become_user, default_exe, success,
-                                                              default_cmd), cmd) is not None)
+    assert re.match("""%s %s -u %s 'echo %s; %s'""" % (pbrun_exe, pbrun_flags, play_context.become_user,
+                                                       success, default_cmd), cmd) is not None
